Updated code with new GUI.
Added/Updated bindings for all GUI elements.
Updated the display to show the interaction info for an interaction
that happens in the game:
    Game win/lose
    Taking damage from a hazard

Fixed combobox to always have an available action selected. It should no
longer be blank ever unless there are no actions to take.

If gameover (win/lose) the take action button is disabled.